Decoder: add frame rate to caps
The decoder doens't set framerate to the capabilities itself and fails
while calling gst_pad_alloc_buffer_and_set_caps because the caps from
outbuffer, whereever it gets it from, is not fixed. Hence we add the
frame_rate_de and frame_rate_nu to the decoder struct and add it
manually to the decoder pad.

since we are at it. I broke a little with Lindent to make the caps struct
somehow readble

Signed-off-by: Luotao Fu <l.fu@pengutronix.de>