vpu kernel module: initialize CMD_SET_FRAME_SOURCE_BUF_STRIDE
This register hasn't been handled in the older Freescale code
at all. The newer code sets this unconditionally on both
i.MX27 and i.MX51/53. Now the question if what meaning it
has on i.MX27? None? the same? some other?

Nobody know. As it doesn't seem to be needed on i.MX27 put
this in v2 specific code.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>