vpu module decoder: fix race condition
When we start a frame and the interrupt handler detects that
there is not enough data in the fifo to decode a frame it
sets the instance on hold. This is fine, but it fails when
in the mean time data arrived in the fifo. So, use a variable
to track this.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>